Encode session id as hex instead of base64
The base64 encoding had to be reencoded again anyway, because it
contains invalid characters for a cookie. Thus, using hex
encoding saves one step and makes more sense.

With this commit, `Cargo.toml` refers to a local path. But that's
fine since I'm working alone for now ^_^
It will be changed back once [1] is merged.

[1]: https://github.com/KokaKiwi/rust-hex/pull/11